refactor(review): split model presentation helpers
